requirements.txt
keybert
fastapi
mysql-connector-python